chore(`generator-nef-library`): move validators to its own file
For easier testing
